+++ Runtime/Core/AbstractCoreAnimation.cs
@@ -9,6 +9,17 @@ namespace ActionCode.AnimationSystem
     {
         [Tooltip("If enabled, animation will play even if Time.deltaTime = 0")]
         public bool useUnscaledTime;
+        [SerializeField, Tooltip("The animation speed."), Min(0f)]
+        private float speed = 1f;
+
+        /// <summary>
+        /// The animation speed.
+        /// </summary>
+        public float Speed
+        {
+            get => speed;
+            set => speed = Mathf.Max(value, 0f);
+        }
 
         /// <summary>
         /// The animation current time.
@@ -37,11 +48,18 @@ namespace ActionCode.AnimationSystem
             if (HasCurveFinished(curve, CurrentTime)) Stop();
         }
 
+        protected void CheckStopCondition(Vector3Curve curve)
+        {
+            if (curve.HasCurveFinished(CurrentTime)) Stop();
+        }
+
         private float GetDeltaTime() => useUnscaledTime ? Time.unscaledDeltaTime : Time.deltaTime;
+
+        public static bool IsLoop(WrapMode mode) => mode is WrapMode.Loop or WrapMode.PingPong;
 
         public static bool HasCurveFinished(AnimationCurve curve, float currentTime)
         {
-            var isLoop = curve.postWrapMode is WrapMode.Loop or WrapMode.PingPong;
+            var isLoop = IsLoop(curve.postWrapMode);
             if (isLoop) return false;
             return currentTime >= GetDuration(curve);
         }