+++ Runtime/Core/Transform/RotationAnimation.cs
@@ -8,13 +8,15 @@ namespace ActionCode.AnimationSystem
     [AddComponentMenu("Animation/Transform/Rotation")]
     public sealed class RotationAnimation : AbstractCoreAnimation
     {
+        [Tooltip("Whether to rotate locally or relative to the Scene in world space.")]
         public Space relation = Space.Self;
-        public Vector3 speed = Vector3.up * 15f;
+        [Tooltip("The axis used to rotate.")]
+        public Vector3 axis = Vector3.up;
 
         protected override void UpdateAnimation(float time)
         {
             base.UpdateAnimation(time);
-            var velocity = speed * time;
+            var velocity = Speed * time * axis;
             transform.Rotate(velocity, relation);
         }
     }